backend-project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
cachetools==5.5.0
certifi==2024.8.30
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
constantly==23.10.4
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
frozenlist==1.6.0
fuzzywuzzy==0.18.0
pymysql==1.1.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Levenshtein==0.27.1
lxml==5.4.0
msgpack==1.1.0
multidict==6.4.3
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
proto-plus==1.25.0
protobuf==5.29.1
python-docx==4.13.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==1.10.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
RapidFuzz==3.13.0
redis==6.0.0
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.2
yarl==1.20.0
zope.interface==7.2
